requirements.txt
# Core
openai==1.35.7

# Document Loaders
pymupdf==1.24.5
python-docx==1.1.2
beautifulsoup4==4.12.3
requests==2.32.3

# Language Detection
langdetect==1.0.9

# API Layer
fastapi==0.111.1
uvicorn==0.30.1
python-multipart==0.0.9

# Database
supabase==2.5.1

# Utilities
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
structlog==24.2.0
